Move 'Polygon'-related strategies
