+++ tests/shaped_tests/polygon_tests/test_area.py
@@ -1,7 +1,7 @@
 from hypothesis import given
 
 from gon.shaped import Polygon
-from tests import strategies
+from . import strategies
 
 
 @given(strategies.polygons)
