+++ tests/shaped_tests/polygon_tests/test_hash.py
@@ -1,8 +1,8 @@
 from hypothesis import given
 
 from gon.shaped import Polygon
-from tests import strategies
 from tests.utils import equivalence
+from . import strategies
 
 
 @given(strategies.polygons)
