+++ tests/strategies/__init__.py
@@ -4,7 +4,4 @@ from .base import (points,
                    scalars_to_points)
 from .linear import segment_to_scalars
 from .shaped import (invalid_vertices,
-                     polygons,
-                     polygons_with_points,
-                     triangles,
-                     triangles_vertices)
+                     to_non_triangle_vertices_base)
